+++ frontend/src/setupTests.js
@@ -1,4 +1,5 @@
 import { configure } from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
+import "./test_utils";
 
 configure({ adapter: new Adapter() });
